refactor(agents): extract transport-proxy recovery from lifecycle_ops
lifecycle_ops.rs is the single most fix-churned file in the codebase and
the largest module in agent_task_lifecycle (3708 LOC). The controller
transport-proxy recovery cluster (TransportProxyRecovery, recover_transport_proxy,
recover_terminal_transport_proxy_evidence, resume/reconcile/bind helpers, and the
transport-proxy predicates) is a cohesive partial-failure surface that attracts
recurring Lab-handoff fixes.

Extract it verbatim into lifecycle_transport_proxy.rs so its reconnect/resume/replay
invariants are inspectable in isolation. Behavior-preserving pure move: the module
opens with use super::* and is re-exported through mod.rs, so all external
agent_task_lifecycle:: paths resolve unchanged. is_transport_proxy widened to
pub(crate) for the one remaining lifecycle_ops caller.

lifecycle_ops.rs: 3708 -> 3420 LOC.
