+++ crates/homeboy-agents/src/agent_task_lifecycle/mod.rs
@@ -39,6 +39,7 @@ mod failure_recording;
 mod health;
 mod lifecycle_ops;
 mod lifecycle_record_ops;
+mod lifecycle_transport_proxy;
 mod records;
 pub mod runner_continuation;
 
@@ -49,6 +50,7 @@ pub use health::*;
 pub use homeboy_core::controller_runtime::ControllerRuntimePruneResult;
 pub use lifecycle_ops::*;
 pub use lifecycle_record_ops::cook_attempt_run_id;
+pub use lifecycle_transport_proxy::*;
 pub use records::*;
 #[cfg(any(test, feature = "test-support"))]
 pub use runner_continuation::{
